Backend/requirements.txt
# Web framework
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.26

# API documentation
scalar-fastapi==1.8.2
Jinja2==3.1.6

# Database - SQL Server
sqlalchemy==2.0.49
pyodbc==5.3.0

# Data validation
pydantic==2.11.9

# HTTP client (async - WhatsApp)
httpx==0.28.1

# Environment variables
python-dotenv==1.1.0

# AI providers
openai==2.1.0
google-generativeai==0.8.6
anthropic==0.96.0
